increased default socket timeout to 100 ms
After fleshing out more of the connection timeouts,
it seems that random timeouts were occuring on the first
attempt to connect, while subsequent calls were reliable.
These turned out to be ARPs for the server or client,
adding additional delay. Picking a more sensible value
now that the timeouts are actually being obeyed.