ease up on the server timeout checks
testServerConnection() was added for long waiting calls
to differentiate between server down and lengthy calls.
It was probably too harsh and could lead to throwing timeout
when the server was just bogged down or on a slower embedded machine.
This should really be implemented with a in-application layer
keep-alive (master), but until then this will still help to exit
if a server crashes/link down earlier than the full timeout.